Move imod.mf6.statusinfo.py to imod.common.statusinfo.py
