+++ imod/common/interfaces/imodel.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Optional, Tuple
 
 from imod.common.interfaces.idict import IDict
-from imod.mf6.statusinfo import StatusInfoBase
+from imod.common.statusinfo import StatusInfoBase
 from imod.typing import GridDataArray
