+++ imod/mf6/utilities/regrid.py
@@ -15,11 +15,11 @@ from imod.common.interfaces.ipackage import IPackage
 from imod.common.interfaces.ipointdatapackage import IPointDataPackage
 from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.common.interfaces.isimulation import ISimulation
+from imod.common.statusinfo import NestedStatusInfo
 from imod.mf6.auxiliary_variables import (
     expand_transient_auxiliary_variables,
     remove_expanded_auxiliary_variables_from_dataset,
 )
-from imod.mf6.statusinfo import NestedStatusInfo
 from imod.mf6.utilities.clip import clip_by_grid
 from imod.mf6.utilities.package import _is_valid
 from imod.schemata import ValidationError
